rename getPagination to usePagination
